+++ src/components/project-preference-card.tsx
