+++ src/lib/validations/board.ts
@@ -3,3 +3,10 @@ type Id = string | number;
 export type BoardColumn = { id: Id; displayName: string };
 
 export type ProjectPreference = { id: Id; columnId: Id; title: string };
+
+export type SortablePreference = {
+  id: string;
+  title: string;
+  rank: number;
+  changed?: boolean;
+};
